# [PERF] stock: reduce forecast info query count