@zaha-odoo 2024-06-03
This is a resubmission of https://github.com/odoo/odoo/pull/166784 due to it having a bad branch name.
Please review the discussion there as it may answer some questions.
Thank you :)