@zaha-odoo 2024-06-03
Amended with formatting/style fixes