@zaha-odoo 2024-06-05
See also opw-3958462 which exhibits this issue